requirements.txt
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
choicelib==0.1.5
exceptiongroup==1.1.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pydantic==1.10.12
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==2.0.4
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
yarl==1.9.2
